Add successful layers parameter to the buildFailure hook
It may be useful for a cloud service to store all of the layers which
did succeed when a build fails, and before this commit that was not
possible without doing some log parsing.

This commit adds a `layers` paramter to the buildFailure hook, which is
all of the successful layers.

Also added tests.

Change-type: minor
Signed-off-by: Cameron Diver <cameron@resin.io>